Backend/recitation-compare/files/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
librosa
torch
transformers